+++ core/src/main/java/cucumber/runtime/AmbiguousStepDefinitionsException.java
@@ -6,11 +6,11 @@ public class AmbiguousStepDefinitionsException extends CucumberException {
     private final List<StepDefinitionMatch> matches;
 
     public AmbiguousStepDefinitionsException(List<StepDefinitionMatch> matches) {
-        super(getMessage(matches));
+        super(createMessage(matches));
         this.matches = matches;
     }
 
-    private static String getMessage(List<StepDefinitionMatch> matches) {
+    private static String createMessage(List<StepDefinitionMatch> matches) {
         StringBuilder msg = new StringBuilder();
         msg.append(matches.get(0).getStepLocation()).append(" matches more than one step definition:\n");
         for (StepDefinitionMatch match : matches) {