+++ core/src/test/java/cucumber/runtime/RuntimeGlueTest.java
@@ -1,0 +1,32 @@
+package cucumber.runtime;
+
+import cucumber.runtime.converters.LocalizedXStreams;
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.fail;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+public class RuntimeGlueTest {
+    @Test
+    public void throws_duplicate_error_on_dupe_stepdefs() {
+        RuntimeGlue glue = new RuntimeGlue(new UndefinedStepsTracker(), new LocalizedXStreams(Thread.currentThread().getContextClassLoader()));
+
+        StepDefinition a = mock(StepDefinition.class);
+        when(a.getPattern()).thenReturn("hello");
+        when(a.getLocation()).thenReturn("foo.bf:10");
+        glue.addStepDefinition(a);
+
+        StepDefinition b = mock(StepDefinition.class);
+        when(b.getPattern()).thenReturn("hello");
+        when(b.getLocation()).thenReturn("bar.bf:90");
+        try {
+            glue.addStepDefinition(b);
+            fail("should have failed");
+        } catch (DuplicateStepDefinitionException expected) {
+            assertEquals("Duplicate step definitions in foo.bf:10 and bar.bf:90", expected.getMessage());
+        }
+    }
+
+}